refactor: split logger config and service
