+++ Cargo.toml
@@ -20,6 +20,7 @@ members = [
     # Members are ordered by dependencies. Crates at top has fewer dependencies.
     "util/build-info",
     "util/logger",
+    "util/logger-config",
     "util/logger-service",
     "util/memory-tracker",
     "util",
