+++ util/logger-config/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "ckb-logger-config"
+version = "0.36.0-pre"
+license = "MIT"
+authors = ["Nervos <dev@nervos.org>"]
+edition = "2018"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
