+++ util/logger-service/Cargo.toml
@@ -7,13 +7,13 @@ edition = "2018"
 
 [dependencies]
 ckb-util = { path = ".." }
+ckb-logger-config = { path = "../logger-config" }
 ansi_term = "0.12"
 log = "0.4"
 env_logger = "0.6"
 lazy_static = "1.3"
 regex = "1.1.6"
 chrono = "0.4"
-serde = { version = "1.0", features = ["derive"] }
 crossbeam-channel = "0.3"
 backtrace = "0.3"
 sentry = "0.16.0"
