+++ util/logger-service/src/lib.rs
@@ -1,17 +1,18 @@
 use ansi_term::Colour;
 use backtrace::Backtrace;
 use chrono::prelude::{DateTime, Local};
-use ckb_util::{Mutex, RwLock};
 use crossbeam_channel::unbounded;
 use env_logger::filter::{Builder, Filter};
 use lazy_static::lazy_static;
 use log::{LevelFilter, Log, Metadata, Record, SetLoggerError};
 use regex::Regex;
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::io::Write;
 use std::path::PathBuf;
 use std::{fs, panic, process, sync, thread};
+
+use ckb_logger_config::Config;
+use ckb_util::{strings, Mutex, RwLock};
 
 lazy_static! {
     static ref CONTROL_HANDLE: sync::Arc<RwLock<Option<crossbeam_channel::Sender<Message>>>> =
@@ -364,23 +365,7 @@ impl Logger {
     }
 
     pub fn check_extra_logger_name(name: &str) -> Result<(), String> {
-        if name.is_empty() {
-            return Err("the name of extra shouldn't be empty".to_owned());
-        }
-        match Regex::new(r"^[0-9a-zA-Z_-]+$") {
-            Ok(re) => {
-                if !re.is_match(&name) {
-                    return Err(format!(
-                        "invaild extra logger name \"{}\", only \"0-9a-zA-Z_-\" are allowed",
-                        name
-                    ));
-                }
-            }
-            Err(err) => {
-                return Err(format!("failed to check the name of extra logger: {}", err));
-            }
-        }
-        Ok(())
+        strings::check_if_identifier_is_valid(name)
     }
 
     pub fn update_extra_logger(name: String, filter_str: String) -> Result<(), String> {
@@ -392,41 +377,6 @@ impl Logger {
     pub fn remove_extra_logger(name: String) -> Result<(), String> {
         let message = Message::RemoveExtraLogger(name);
         Self::send_message(message)
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct Config {
-    pub filter: Option<String>,
-    pub color: bool,
-    #[serde(skip)]
-    pub file: PathBuf,
-    #[serde(skip)]
-    pub log_dir: PathBuf,
-    pub log_to_file: bool,
-    pub log_to_stdout: bool,
-    pub emit_sentry_breadcrumbs: Option<bool>,
-    #[serde(default)]
-    pub extra: HashMap<String, ExtraLoggerConfig>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct ExtraLoggerConfig {
-    pub filter: String,
-}
-
-impl Default for Config {
-    fn default() -> Self {
-        Config {
-            filter: None,
-            color: !cfg!(windows),
-            file: Default::default(),
-            log_dir: Default::default(),
-            log_to_file: false,
-            log_to_stdout: true,
-            emit_sentry_breadcrumbs: None,
-            extra: Default::default(),
-        }
     }
 }
